Feature/get type back

See merge request s09-fintech-finance-sub2/S09P22A207!41

+++ server/src/main/java/com/galaxy/memorit/category/infrastructure/persistence/repository/BaseCategoryRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 
 import com.galaxy.memorit.category.infrastructure.persistence.entity.BaseCategoryEntity;
 @Repository
-public interface BaseCategoryRepository extends JpaRepository<BaseCategoryEntity, Integer> {
+public interface BaseCategoryRepository extends JpaRepository<BaseCategoryEntity, Long> {
 }
